pyproject.toml
[project]
name = "org-manager"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "email-validator>=2.1",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]


